mofified to place the files in any place of harddrive

+++ IntroAlgoApp.py
@@ -4,7 +4,7 @@ from bs4 import BeautifulSoup
 
 web_page = 'https://ita.skanev.com'
 r = requests.get(web_page)
-htmlSoup = BeautifulSoup(r.text,"html.parser")
+htmlSoup = BeautifulSoup(r.text, "html.parser")
 a = htmlSoup.find_all('a')
 
 # First 3 links are intro,exercise,about
@@ -12,14 +12,15 @@ a = htmlSoup.find_all('a')
 ref = []
 for x in a:
     ref.append(str(x))
-
-# print(ref[3])
-
-# inputex=input()
-path = 'C:\\Users\\nEW u\\Desktop\\study\\4th sem\\Design &Analysis of Algorithms'
+#'C:\\Users\\nEW u\\Desktop\\study\\4th sem\\Design &Analysis of Algorithms'
+path = input('Enter the path where to create a new folder :')
 os.chdir(path)
-os.makedirs('Answers-Introduction-to-Algorithms')
-# path=path+'\\Answers-Introduction-to-Algorithms'
+path=path+'\\Answers-Introduction-to-Algorithms'
+if not os.path.exists(path):
+    os.makedirs('Answers-Introduction-to-Algorithms')
+
+os.chdir(path)
+
 extension = ''
 for line in range(len(ref)):
     fragment = ref[line].split('"')
@@ -28,25 +29,16 @@ for line in range(len(ref)):
     req = requests.get(r.url+extension)
     try :
         req.raise_for_status()
-
         filename = name[1].strip()+'.html'
-        file_object = open(filename,'wb')
-        for chunk in req.iter_content(100000):
-            file_object.write(chunk)
-        file_object.close()
-        print(extension)
+        if not os.path.isfile(filename):
+            file_object = open(filename, 'wb')
+            for chunk in req.iter_content(100000):
+                file_object.write(chunk)
+            file_object.close()
+        print("currently copying", extension)
     except Exception as exc:
         pass
 
-    '''for index in range(9,22,1):
-		extension=extension+ref[line][index]
-	print(extension)
-    extension=''
-    '''
-
-#extensions for problems differ normal
-
-#C:\Users\nEW u\Desktop\study\4th sem\Design &Analysis of Algorithms
